fix: address cycle 7 review blockers — dtype preservation, memory release, and test rigor
1. _load_model_from_artifact: preserve per-key artifact dtype instead of
   coercing all tensors to a single storage_dtype. Mixed-dtype artifacts
   (e.g. float32+float16) now load correctly.

2. streaming_evaluation_to_sink: explicitly release merged_cpu after
   write_fn in both the normal and OOM-retry paths so previous chunks
   are not resident while the next is built/evaluated.

3. test_full_mode_succeeds_when_inmemory_sink_construction_fails: now
   monkeypatches both chunked_evaluation AND evaluate_affected_group to
   raise, proving full mode uses streaming_evaluation_to_sink exclusively.

4. test_event_order_interleaved_keys: asserts base_only key is written,
   verifies base writes precede group 1 evaluation, carries
   ac-base-weight-bounded-copying annotation.

5. test_failure_during_eval_aborts_sink: real runtime test with two
   groups — first group writes tensors, second group raises. Verifies
   abort/no-artifact/no-cache-hit/no-payload chain.

6. test_pre_populated_patch_cache_not_read_by_full_mode: uses sentinel
   zeros vs fresh 42.0 values to prove sentinel was not read. New
   test_full_artifact_not_treated_as_patch_tensor_payload for cross-mode
   isolation.

7. Dynamic VRAM tests: _run_with_vram_state helper installs mock
   comfy.model_management with distinct vram_state values and tracks
   set_vram_state calls. Each test simulates a different memory mode
   and asserts no mode-changing API was called.

8. TestMixedDtypeArtifactPreservation: new test verifying float32/float16
   artifact tensors retain their dtypes after loading.

Task: @task-full-model-streaming-artifact-cache
Spec: @streaming-full-model-materialization

+++ nodes/exit.py
@@ -345,10 +345,12 @@ def _load_model_from_artifact(
     # Load artifact weights directly into the clone's model state dict.
     # This makes the weights model-owned (Comfy-managed) rather than
     # patch-owned (always resident).
+    # Preserve each tensor's artifact dtype — do NOT coerce to a single
+    # global storage_dtype, which would destroy mixed-dtype artifacts.
     artifact_state: dict[str, torch.Tensor] = {}
     with safe_open(save_path, framework="pt", device="cpu") as f:
         for key in f.keys():
-            artifact_state[key] = f.get_tensor(key).to(dtype=storage_dtype)
+            artifact_state[key] = f.get_tensor(key)
 
     # Update the clone's underlying model weights with artifact data.
     # Try load_state_dict (real nn.Module) first, then fall back to
